+++ Projet_Site_e_commerce/src/main/java/fr/adaming/dao/ICategorieDao.java
@@ -9,6 +9,6 @@ public interface ICategorieDao {
 	public Categorie ajouterCategorie(Categorie c);
 	public Categorie supprimerCategorie(Categorie c);
 	public Categorie modifierCategorie(Categorie c);
-	public Categorie getCategorieById(int id);
+	public Categorie getCategorieById(long id);
 	public List<Categorie> getAllCategories();
 }
